+++ src/main/java/com/mageddo/dnsproxyserver/config/dataprovider/ConfigDAOJson.java
@@ -2,8 +2,6 @@ package com.mageddo.dnsproxyserver.config.dataprovider;
 
 import com.mageddo.dnsproxyserver.config.Config;
 import com.mageddo.dnsproxyserver.config.dataprovider.mapper.ConfigJsonV2Mapper;
-import com.mageddo.utils.Files;
-import com.mageddo.utils.Runtime;
 import com.mageddo.utils.Tests;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -11,7 +9,6 @@ import lombok.extern.slf4j.Slf4j;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.Arrays;
 
 @Slf4j
@@ -26,7 +23,7 @@ public class ConfigDAOJson implements ConfigDAO {
   public Config find() {
     final var workDir = this.configDAOEnv.findRaw().getCurrentPath();
     final var relativeConfigFilePath = this.configDAOCmdArgs.findRaw().getConfigPath();
-    final var configFileAbsolutePath = buildConfigPath(workDir, relativeConfigFilePath);
+    final var configFileAbsolutePath = ConfigPathBuilder.build(workDir, relativeConfigFilePath);
     return this.find(configFileAbsolutePath);
   }
 
@@ -34,29 +31,6 @@ public class ConfigDAOJson implements ConfigDAO {
     final var jsonConfig = JsonConfigs.loadConfig(configPath);
     log.debug("configPath={}", configPath);
     return ConfigJsonV2Mapper.toConfig(jsonConfig, configPath);
-  }
-
-  public static Path buildConfigPath(Path workDir, String configPath) {
-    if (runningInTestsAndNoCustomConfigPath()) {
-      final var file = Files.createTempFileDeleteOnExit("dns-proxy-server-junit", ".json");
-      log.trace("status=runningInTests, usingEmptyFile={}", file);
-      return file;
-    }
-    if (workDir != null) {
-      return workDir
-        .resolve(configPath)
-        .toAbsolutePath()
-        ;
-    }
-    final var confRelativeToCurrDir = Paths
-      .get(configPath)
-      .toAbsolutePath();
-    if (Files.exists(confRelativeToCurrDir)) {
-      return confRelativeToCurrDir;
-    }
-    return Runtime.getRunningDir()
-      .resolve(configPath)
-      .toAbsolutePath();
   }
 
   static boolean runningInTestsAndNoCustomConfigPath() {
